@@ -1,19 +1,7 @@
 import '../entities/inventory_item.dart';
 
-/// Abstract repository for inventory data operations
 abstract class InventoryRepository {
-  /// Fetches all inventory items from the remote server
-  Future<List<InventoryItem>> fetchInventory();
-
-  /// Updates an existing inventory item
-  Future<InventoryItem> updateItem(InventoryItem item);
-
-  /// Deletes an inventory item by ID
-  Future<void> deleteItem(int id);
-
-  /// Adds a new inventory item
-  Future<InventoryItem> addItem(InventoryItem item);
-
-  /// Searches inventory items by query
-  Future<List<InventoryItem>> searchItems(String query);
+  Future<List<InventoryItem>> fetchAll();
+  Future<InventoryItem> updateItem(String id, Map<String, dynamic> updates);
+  Future<void> deleteItem(String id);
 }
