@@ -2,133 +2,33 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'core/theme/app_theme.dart';
-import 'presentation/providers/auth_provider.dart';
-import 'presentation/screens/login_screen.dart';
-import 'presentation/screens/inventory_list_screen.dart';
-import 'presentation/screens/item_detail_screen.dart';
-import 'presentation/screens/camera_screen.dart';
-import 'presentation/screens/barcode_scanner_screen.dart';
+import 'presentation/providers/repository_providers.dart';
+import 'presentation/routes/app_router.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+
   final prefs = await SharedPreferences.getInstance();
-  runApp(ProviderScope(
-    overrides: [
-      sharedPreferencesProvider.overrideWithValue(prefs),
-    ],
-    child: const MyApp(),
-  ));
+
+  runApp(
+    ProviderScope(
+      overrides: [sharedPreferencesProvider.overrideWithValue(prefs)],
+      child: const MainApp(),
+    ),
+  );
 }
 
-class MyApp extends ConsumerStatefulWidget {
-  const MyApp({super.key});
+class MainApp extends ConsumerWidget {
+  const MainApp({super.key});
 
   @override
-  ConsumerState<MyApp> createState() => _MyAppState();
-}
+  Widget build(BuildContext context, WidgetRef ref) {
+    final appRouter = ref.watch(appRouterProvider);
 
-class _MyAppState extends ConsumerState<MyApp> {
-  @override
-  void initState() {
-    super.initState();
-    // Check authentication status on app start
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      ref.read(authProvider.notifier).checkAuthStatus();
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Inventory Nexus',
-      theme: AppTheme.lightTheme,
-      debugShowCheckedModeBanner: false,
-      initialRoute: '/login',
-      onGenerateRoute: _generateRoute,
+    return MaterialApp.router(
+      title: 'Stitch App',
+      theme: AppTheme.light,
+      routerConfig: appRouter,
     );
-  }
-
-  Route<dynamic> _generateRoute(RouteSettings settings) {
-    switch (settings.name) {
-      case '/login':
-        return MaterialPageRoute(
-          builder: (_) => const LoginScreen(),
-          settings: settings,
-        );
-      case '/inventory':
-        return MaterialPageRoute(
-          builder: (_) => const InventoryListScreen(),
-          settings: settings,
-        );
-      case '/inventory/detail':
-        return PageRouteBuilder(
-          settings: settings,
-          pageBuilder: (context, animation, secondaryAnimation) =>
-              const ItemDetailScreen(),
-          transitionsBuilder: (context, animation, secondaryAnimation, child) {
-            const begin = Offset(1.0, 0.0);
-            const end = Offset.zero;
-            const curve = Curves.easeInOut;
-
-            var tween = Tween(begin: begin, end: end).chain(
-              CurveTween(curve: curve),
-            );
-
-            return SlideTransition(
-              position: animation.drive(tween),
-              child: child,
-            );
-          },
-        );
-      case '/camera':
-        return PageRouteBuilder(
-          settings: settings,
-          pageBuilder: (context, animation, secondaryAnimation) => CameraScreen(
-            onPhotoCaptured: (path) {
-              // Handle photo capture
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(content: Text('Photo saved: $path')),
-              );
-            },
-          ),
-          transitionsBuilder: (context, animation, secondaryAnimation, child) {
-            const begin = Offset(0.0, 1.0);
-            const end = Offset.zero;
-            const curve = Curves.easeInOut;
-
-            var tween = Tween(begin: begin, end: end).chain(
-              CurveTween(curve: curve),
-            );
-
-            return SlideTransition(
-              position: animation.drive(tween),
-              child: child,
-            );
-          },
-        );
-      case '/scanner':
-        return PageRouteBuilder(
-          settings: settings,
-          pageBuilder: (context, animation, secondaryAnimation) =>
-              BarcodeScannerScreen(
-            onBarcodeDetected: (value) {
-              // Handle barcode scan
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(content: Text('Barcode scanned: $value')),
-              );
-            },
-          ),
-          transitionsBuilder: (context, animation, secondaryAnimation, child) {
-            return FadeTransition(
-              opacity: animation,
-              child: child,
-            );
-          },
-        );
-      default:
-        return MaterialPageRoute(
-          builder: (_) => const LoginScreen(),
-        );
-    }
   }
 }
