@@ -1,129 +1,29 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../../domain/usecases/auth_usecases.dart';
-import '../../domain/repositories/auth_repository.dart';
-import '../../data/repositories/auth_repository_impl.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import '../../domain/entities/user_session.dart';
+import 'repository_providers.dart';
 
-/// Provider for SharedPreferences instance
-final sharedPreferencesProvider = Provider<SharedPreferences>((ref) {
-  throw UnimplementedError('SharedPreferences must be initialized first');
-});
-
-/// Provider for AuthRepository
-final authRepositoryProvider = Provider<AuthRepository>((ref) {
-  final prefs = ref.watch(sharedPreferencesProvider);
-  return AuthRepositoryImpl(prefs);
-});
-
-/// Provider for LoginUseCase
-final loginUseCaseProvider = Provider<LoginUseCase>((ref) {
-  return LoginUseCase(ref.watch(authRepositoryProvider));
-});
-
-/// Provider for LogoutUseCase
-final logoutUseCaseProvider = Provider<LogoutUseCase>((ref) {
-  return LogoutUseCase(ref.watch(authRepositoryProvider));
-});
-
-/// Provider for CheckAuthUseCase
-final checkAuthUseCaseProvider = Provider<CheckAuthUseCase>((ref) {
-  return CheckAuthUseCase(ref.watch(authRepositoryProvider));
-});
-
-/// State for authentication
-enum AuthStatus { initial, loading, authenticated, unauthenticated }
-
-class AuthState {
-  final AuthStatus status;
-  final String? errorMessage;
-
-  AuthState({required this.status, this.errorMessage});
-
-  AuthState copyWith({
-    AuthStatus? status,
-    String? errorMessage,
-  }) {
-    return AuthState(
-      status: status ?? this.status,
-      errorMessage: errorMessage ?? this.errorMessage,
-    );
-  }
-}
-
-/// Notifier for authentication state
-class AuthNotifier extends StateNotifier<AuthState> {
-  final LoginUseCase _loginUseCase;
-  final LogoutUseCase _logoutUseCase;
-  final CheckAuthUseCase _checkAuthUseCase;
-
-  AuthNotifier({
-    required LoginUseCase loginUseCase,
-    required LogoutUseCase logoutUseCase,
-    required CheckAuthUseCase checkAuthUseCase,
-  })  : _loginUseCase = loginUseCase,
-        _logoutUseCase = logoutUseCase,
-        _checkAuthUseCase = checkAuthUseCase,
-        super(AuthState(status: AuthStatus.initial));
-
-  /// Check if user is authenticated on app startup
-  Future<void> checkAuthStatus() async {
-    state = state.copyWith(status: AuthStatus.loading);
-    try {
-      final isAuthenticated = await _checkAuthUseCase();
-      state = state.copyWith(
-        status: isAuthenticated
-            ? AuthStatus.authenticated
-            : AuthStatus.unauthenticated,
-      );
-    } catch (e) {
-      state = state.copyWith(
-        status: AuthStatus.unauthenticated,
-        errorMessage: e.toString(),
-      );
-    }
+class AuthNotifier extends AsyncNotifier<UserSession?> {
+  @override
+  Future<UserSession?> build() async {
+    final repository = ref.read(authRepositoryProvider);
+    return await repository.getSession();
   }
 
-  /// Login with email and password
-  Future<void> login(String email, String password) async {
-    state = state.copyWith(status: AuthStatus.loading, errorMessage: null);
-    try {
-      final success = await _loginUseCase(email, password);
-      if (success) {
-        state = state.copyWith(status: AuthStatus.authenticated);
-      } else {
-        state = state.copyWith(
-          status: AuthStatus.unauthenticated,
-          errorMessage: 'Invalid credentials',
-        );
-      }
-    } catch (e) {
-      state = state.copyWith(
-        status: AuthStatus.unauthenticated,
-        errorMessage: 'Login failed: ${e.toString()}',
-      );
-    }
+  Future<void> login(String userId, String password) async {
+    state = const AsyncValue.loading();
+    state = await AsyncValue.guard(() async {
+      final repository = ref.read(authRepositoryProvider);
+      return await repository.login(userId, password);
+    });
   }
 
-  /// Logout current user
   Future<void> logout() async {
-    state = state.copyWith(status: AuthStatus.loading);
-    try {
-      await _logoutUseCase();
-      state = state.copyWith(status: AuthStatus.unauthenticated);
-    } catch (e) {
-      state = state.copyWith(
-        status: AuthStatus.authenticated,
-        errorMessage: 'Logout failed: ${e.toString()}',
-      );
-    }
+    final repository = ref.read(authRepositoryProvider);
+    await repository.logout();
+    state = const AsyncValue.data(null);
   }
 }
 
-/// Provider for AuthNotifier
-final authProvider = StateNotifierProvider<AuthNotifier, AuthState>((ref) {
-  return AuthNotifier(
-    loginUseCase: ref.watch(loginUseCaseProvider),
-    logoutUseCase: ref.watch(logoutUseCaseProvider),
-    checkAuthUseCase: ref.watch(checkAuthUseCaseProvider),
-  );
-});
+final authProvider = AsyncNotifierProvider<AuthNotifier, UserSession?>(
+  AuthNotifier.new,
+);
